refactor: SpaceRepository,ChannelFacade 이동
