+++ src/main/java/com/echo/echo/domain/channel/ChannelFacade.java
@@ -2,6 +2,10 @@ package com.echo.echo.domain.channel;
 
 import com.echo.echo.domain.channel.dto.ChannelRequestDto;
 import com.echo.echo.domain.channel.dto.ChannelResponseDto;
+import com.echo.echo.domain.space.entity.Space;
+import com.echo.echo.domain.space.error.SpaceErrorCode;
+import com.echo.echo.domain.space.repository.SpaceRepository;
+import com.echo.echo.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
@@ -15,9 +19,11 @@ import reactor.core.publisher.Mono;
 public class ChannelFacade {
 
     private final ChannelService channelService;
+    private final SpaceRepository spaceRepository;
 
     public Mono<ChannelResponseDto> createChannel(Long spaceId, ChannelRequestDto requestDto) {
-        return channelService.createChannel(spaceId, requestDto);
+        return findSpaceById(spaceId)
+            .flatMap(space -> channelService.createChannel(spaceId, requestDto));
     }
 
     public Flux<ChannelResponseDto> getChannels(Long spaceId) {
@@ -30,5 +36,10 @@ public class ChannelFacade {
 
     public Mono<Void> deleteChannel(Long channelId) {
         return channelService.deleteChannel(channelId);
+    }
+
+    private Mono<Space> findSpaceById(Long spaceId) {
+        return spaceRepository.findById(spaceId)
+            .switchIfEmpty(Mono.error(new CustomException(SpaceErrorCode.SPACE_NOT_FOUND)));
     }
 }
